refactor(template): matches foles file structure
