+++ src/components/templates/MahomesTemplate/MahomesTemplate.js
@@ -3,7 +3,7 @@ import { StyleSheet, View } from 'react-pdf';
 import { generatePDFTemplate } from '../BaseTemplate/BaseTemplate';
 import { ContentColumn } from './ContentColumn';
 import { Heading } from './Heading';
-import { InfoColumn } from './InfoColumn';
+import { InfoColumn } from '../MahomesTemplate/info/InfoColumn';
 import data from '../../../example-json/john_smith.json';
 
 const styles = StyleSheet.create({
