+++ src/components/templates/MahomesTemplate/info/info-styles.js
@@ -1,0 +1,47 @@
+import { StyleSheet } from 'react-pdf';
+
+export const styles = StyleSheet.create({
+  'contact-info--contact-item': {
+    display: 'flex',
+    flexDirection: 'row',
+    padding: '0 0 5px 5px',
+  },
+  'column--info': {
+    backgroundColor: '#EEE',
+    color: 'black',
+    padding: '6px 0 0 12px',
+    width: '35%',
+  },
+  'heading-info--name': {
+    fontSize: 24,
+    fontWeight: 'bold',
+  },
+  'section-info--heading': {
+    fontSize: 16,
+    fontWeight: 'bold',
+    margin: '0 0 8px 0',
+    padding: '0 0 0 0px',
+  },
+  'section-info--heading-second': {
+    fontSize: 16,
+    fontWeight: 'bold',
+    margin: '0 0 8px 0',
+    padding: '15px 0 0 0px',
+  },
+  'section--contact-info': {
+    padding: '0 0 0 0px',
+  },
+  'section--heading-info': {
+    padding: '0 0 0 0px',
+  },
+  'section--skills-info': {
+    padding: '0 0 0 0px',
+  },
+  'section--skills-info--text': {
+    marginBottom: '10px',
+    marginLeft: '10px',
+  },
+  'heading--location-item': {
+    padding: '0 0 20px 5px',
+  },
+});
